Split storage replication core tests
