+++ .pm/tasks/task_082d3f35aee2436c97ba2a629ec8437a.execution.md
@@ -169,3 +169,12 @@ Example:
 - Expected Result: The committed hotfix remains covered by focused regression, storage replication suite, formatting, and whitespace checks while PR/CI provides the next authoritative gate.
 - Actual Result: PASS before commit 74e247d3e: focused regression passed 1/1; storage_replication filter passed 34/34; fmt --check passed; git diff --check passed.
 - Blocker / Next Action: No local verification blocker; push a PR branch, open PR, watch required checks/comments, merge when acceptable, then trigger Testnet Packages and deploy.
+
+## 2026-06-20 20:11:52 CST / tpm
+- 完成内容: PR #545 REQUIRED-GATE FAILURE TRIAGED AND FIXED. GitHub Rust required-gate failed in `./scripts/check-rust-file-size.sh` because adding the pending-retention regression pushed `crates/oasis7_node/src/tests_storage_replication.rs` to 1260 lines, above the 1200-line test-file limit.
+- 遗留事项: Push the structural fix and watch PR #545 checks again.
+- Action: Moved the adjacent non-proposer/proposer local replication tests into new sibling module `crates/oasis7_node/src/tests_storage_replication_core.rs` and added the module include from `tests_storage_replication.rs`. Production logic and assertions are unchanged; the parent file is now 987 lines and the new core test module is 276 lines.
+- Validation Command: ./scripts/check-rust-file-size.sh; ./scripts/cargo-dev.sh test -p oasis7_node non_proposer_missing_remote_replication_retains_pending_commit -- --nocapture; ./scripts/cargo-dev.sh fmt --check; ./scripts/cargo-dev.sh test -p oasis7_node storage_replication -- --nocapture; git diff --check.
+- Expected Result: File-size gate passes; focused regression remains green from its new module path; storage_replication suite remains green; formatting and whitespace checks pass.
+- Actual Result: PASS. check-rust-file-size reported zero oversized files. Focused regression passed 1/1 as `tests::storage_replication_core_tests::non_proposer_missing_remote_replication_retains_pending_commit`. `fmt --check` passed. storage_replication filter passed 37/37. `git diff --check` passed.
+- Blocker / Next Action: No local blocker; commit and push PR #545 fix, then continue CI watch and merge/package/deploy path.
